- Updates to the DBSystem, MySQL ports can now be defined. A server port has also been added, but it is currently not hooked up. - ServerList system has been implemented. You can define the server list within MySQL, sql provided. Know that Groupnames AND ServerNames MUST be unique, two groups CANNOT have the same servername within it. - Improvements to the AlefPacketInterface, there is one last improvement that can be done related to gathering an array of packets to be processed/written. After that it should be final. - Changed some of the startup structuring to give better focus on any errors that may arise during initialization. - Start of the UserLogin system, currently disabled. - Lots of cosmetic changes.

+++ AlefCommon/AlefSocket.h
@@ -41,7 +41,7 @@ public:
 			packet->Resize(packetSize);
 
 			for (UInt32 i = 0; i < (packetSize >> 3); i++)
-				blowfish_encrypt(cryptoSession->serverCtx, (UInt32*)(packet->getBuffer() + i * 8), (UInt32*)(packet->getBuffer() + i * 8 + 4));
+				blowfish_encrypt(cryptoSession->serverCtx, (UInt32*)(packet->getBuffer() + i * (__int64)8), (UInt32*)(packet->getBuffer() + i * (__int64)8 + (__int64)4));
 
 			packet->Resize(packetSize + 8);
 			memmove(packet->getBuffer() + 7, packet->getBuffer(), packetSize);
@@ -49,8 +49,6 @@ public:
 			packet->getBuffer()[packetSize + 7] = 0xAF; //Guard Footer Byte
 			*(UInt16*)(packet->getBuffer() + 1) = ((UInt16)packet->getSize()); //Encrypted Size
 			*(UInt32*)(packet->getBuffer() + 3) = ++cryptoSession->packetCounter;
-			
-			//return this->sendBytes(packet->getBuffer(), packet->getSize());
 		}
 		
 		return this->sendBytes(packet->getBuffer(), packet->getSize());
